requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.4
django-allauth==0.61.1
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-extensions==3.2.3
djangorestframework==3.15.1
idna==3.6
oauthlib==3.2.2
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
sqlparse==0.4.4
urllib3==2.2.1
